Use DATABASE_URL instead of POSTGRES_* in docker-compose.yml
We had `POSTGRES_DB`, `POSTGRES_USER` and `POSTGRES_PASSWORD` in our
`docker-compose.yml` but they are not actually used in the code.
Instead we use the `dj_database_url` module, which takes a default url
but can be overridden by providing an environment variable called
`DATABASE_URL`.  I changed the test database host, password and database
name to show this works in the github test action.