Opportunity: add organizers ManyToMany relation
One or more organizers can manage an Opportunity. Use a ManyToMany
relation for now because it's not clear that there can be only one
Organizer per Opportunity.

To associate an Organizer with an Opportunity via the REST API, create
or patch an Opportunity with the organizers field set to [organizer.pk].

When retrieving an Opportunity the organizers field is a list of
Organizer primary keys. This can be changed to more RESTful hyperlinks
in the future, if necessary.

This commit adds coalesce/organizers/test/factories.py so
coalesce/opportunities/test/test_views.py can easily create an Organizer
object for testing.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>