Connect OpportunityDetailsPage to REST API
Fetch the Opportunity object from the REST API instead of using
hardcoded data.

Note that the REST API returns slightly different fields than the
frontend expects. Some of the fields are absent and some of the types
are different. Therefore it's necessary to adjust the template slightly
to display the fields as produced by the REST API.

Since the REST API request may take some time, add a spinner informing
the user that data is being loaded.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>